+++ azure-storage-blob/src/Blob/Models/CreatePageBlobOptions.php
@@ -24,10 +24,8 @@
 
 namespace MicrosoftAzure\Storage\Blob\Models;
 
-use MicrosoftAzure\Storage\Common\Models\TransactionalMD5Trait;
-
 /**
- * Optional parameters for createPageBlobFromContent.
+ * Optional parameters for CreatePageBlob.
  *
  * @category  Microsoft
  * @package   MicrosoftAzure\Storage\Blob\Models
@@ -36,7 +34,7 @@ use MicrosoftAzure\Storage\Common\Models\TransactionalMD5Trait;
  * @license   https://github.com/azure/azure-storage-php/LICENSE
  * @link      https://github.com/azure/azure-storage-php
  */
-class CreatePageBlobFromContentOptions extends CreatePageBlobOptions
+class CreatePageBlobOptions extends CreateBlobOptions
 {
-    use TransactionalMD5Trait;
+    use AccessTierTrait;
 }